+++ crates/web-client/test/key.test.ts
@@ -5,9 +5,7 @@ test.describe("signature", () => {
   test("should produce a valid signature", async ({ page }) => {
     const isValid = await page.evaluate(() => {
       const secretKey = window.SecretKey.withRng();
-      const message = new window.Word(
-        new BigUint64Array([BigInt(1), BigInt(2), BigInt(3), BigInt(4)])
-      );
+      const message = new window.Word(new BigUint64Array([1n, 2n, 3n, 4n]));
       const signature = secretKey.sign(message);
       const isValid = secretKey.publicKey().verify(message, signature);
 
@@ -19,11 +17,9 @@ test.describe("signature", () => {
   test("should not verify the wrong message", async ({ page }) => {
     const isValid = await page.evaluate(() => {
       const secretKey = window.SecretKey.withRng();
-      const message = new window.Word(
-        new BigUint64Array([BigInt(1), BigInt(2), BigInt(3), BigInt(4)])
-      );
+      const message = new window.Word(new BigUint64Array([1n, 2n, 3n, 4n]));
       const wrongMessage = new window.Word(
-        new BigUint64Array([BigInt(5), BigInt(6), BigInt(7), BigInt(8)])
+        new BigUint64Array([5n, 6n, 7n, 8n])
       );
       const signature = secretKey.sign(message);
       const isValid = secretKey.publicKey().verify(wrongMessage, signature);
@@ -38,9 +34,7 @@ test.describe("signature", () => {
   }) => {
     const isValid = await page.evaluate(() => {
       const secretKey = window.SecretKey.withRng();
-      const message = new window.Word(
-        new BigUint64Array([BigInt(1), BigInt(2), BigInt(3), BigInt(4)])
-      );
+      const message = new window.Word(new BigUint64Array([1n, 2n, 3n, 4n]));
       const signature = secretKey.sign(message);
       const differentSecretKey = window.SecretKey.withRng();
       const isValid = differentSecretKey.publicKey().verify(message, signature);
@@ -55,9 +49,7 @@ test.describe("signature", () => {
   }) => {
     const isValid = await page.evaluate(() => {
       const secretKey = window.SecretKey.withRng();
-      const message = new window.Word(
-        new BigUint64Array([BigInt(1), BigInt(2), BigInt(3), BigInt(4)])
-      );
+      const message = new window.Word(new BigUint64Array([1n, 2n, 3n, 4n]));
       const signature = secretKey.sign(message);
       const serializedSignature = signature.serialize();
       const deserializedSignature =
@@ -109,5 +101,50 @@ test.describe("secret key", () => {
       );
     });
     expect(isValid).toEqual(true);
+  });
+});
+
+describe("signing inputs", () => {
+  it("should be able to sign and verify an arbitrary array of felts", async () => {
+    const { isValid, isValidOther } = await testingPage.evaluate(() => {
+      const secretKey = window.SecretKey.withRng();
+      const otherSecretKey = window.SecretKey.withRng();
+      const message = Array.from(
+        { length: 128 },
+        (_, i) => new window.Felt(BigInt(i))
+      );
+      const signingInputs = window.SigningInputs.newArbitrary(message);
+      const signature = secretKey.signData(signingInputs);
+      const isValid = secretKey
+        .publicKey()
+        .verifyData(signingInputs, signature);
+      const isValidOther = otherSecretKey
+        .publicKey()
+        .verifyData(signingInputs, signature);
+
+      return { isValid, isValidOther };
+    });
+    expect(isValid).to.be.true;
+    expect(isValidOther).to.be.false;
+  });
+
+  it("should be able to sign and verify a blind word", async () => {
+    const { isValid, isValidOther } = await testingPage.evaluate(() => {
+      const secretKey = window.SecretKey.withRng();
+      const otherSecretKey = window.SecretKey.withRng();
+      const message = new window.Word(new BigUint64Array([1n, 2n, 3n, 4n]));
+      const signingInputs = window.SigningInputs.newBlind(message);
+      const signature = secretKey.signData(signingInputs);
+      const isValid = secretKey
+        .publicKey()
+        .verifyData(signingInputs, signature);
+      const isValidOther = otherSecretKey
+        .publicKey()
+        .verifyData(signingInputs, signature);
+
+      return { isValid, isValidOther };
+    });
+    expect(isValid).to.be.true;
+    expect(isValidOther).to.be.false;
   });
 });